refactor: Drop the use of `typing.TYPE_CHECKING`
Given we're not using a static type checker in our CI, the imports
protected by TYPE_CHECKING have gotten too out-of-date to be useful.
So, let's try getting rid of TYPE_CHECKING, and deal with circular
import some other way.

This also fixes sphinx audodoc problems associated with TYPE_CHECKING.

Other than monkeying about with the import, the only change I've made
here is to re-import `UpdatedableNode/Group` and `RemoteNode` up a level
in `alpenhorn.daemon` to make importing these easier.  The
pattern_importer extension demonstrates that third-parties are going
to want to have access to these.
