+++ alpenhorn/daemon/__init__.py
@@ -2,3 +2,7 @@
 
 from .. import __version__
 from .entry import entry
+
+# These classes are used by extensions, so let's import them into the daemon
+# base
+from .update import UpdateableNode, UpdateableGroup, RemoteNode
