+++ alpenhorn/io/lustrehsm.py
@@ -15,28 +15,22 @@ This module provides:
 from __future__ import annotations
 
 import logging
+import os
 import pathlib
 import time
-from typing import IO, TYPE_CHECKING
+from collections.abc import Generator, Hashable
+from typing import IO
 
 import peewee as pw
 
 from ..common.util import pretty_bytes, pretty_deltat
+from ..daemon import UpdateableGroup, UpdateableNode
 from ..daemon.querywalker import QueryWalker
-from ..db import ArchiveFileCopy, utcnow
-from ..scheduler import Task
+from ..db import ArchiveFile, ArchiveFileCopy, ArchiveFileCopyRequest, utcnow
+from ..scheduler import FairMultiFIFOQueue, Task
 from .base import BaseNodeRemote
 from .default import DefaultGroupIO
 from .lustrequota import LustreQuotaNodeIO
-
-if TYPE_CHECKING:
-    import os
-    from collections.abc import Generator, Hashable
-
-    from ..db import ArchiveFile, ArchiveFileCopyRequest
-    from ..scheduler import FairMultiFIFOQueue
-    from ..service.update import UpdateableGroup, UpdateableNode
-del TYPE_CHECKING
 
 log = logging.getLogger(__name__)
