+++ alpenhorn/io/transport.py
@@ -3,16 +3,11 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
+import pathlib
 
+from ..daemon import UpdateableNode
+from ..db import ArchiveFileCopyRequest
 from .default import DefaultGroupIO
-
-if TYPE_CHECKING:
-    import pathlib
-
-    from ..archive import ArchiveFileCopyRequest
-    from ..update import UpdateableNode
-del TYPE_CHECKING
 
 log = logging.getLogger(__name__)
